Bugfix and playing with sample "programs".
The "programs" are specified as a convention mapping forms to integers
and instructions.  The instructions are the six fundamental
value-preserving transforms:

     * <-> ((*))
     O <-> O*
  A(B) <-> A(AB)

These are sufficient to produce the rest of the systems. Having the
primitive instructions of our language be these implies that every
program of our "machine language" is a valid logical statement.