On the track.
So logpy can be used to recognize an expression and transform it into an
equivalent expression.  I have the strong feeling that I am missing
something obvious here.  Like, I should be able to craft a version of
transform() that can be used directly with run(), I am almost getting it
I think. Or not.

Now that an equivalence can be used to transform an expression, we can
apply them to existing forms to create new ones, and the little explorer
program in lof.py can generate new forms too, which can be fed back to
generate new forms.

It is the basis for a logic solver thingy. (I'm way out of my depth
here.) If I'm not totally bonkers I think things like the Church encoding
can be rendered fairly directly in these forms.  Where will it all lead?

:)